requirements.txt
Django==5.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
dj-rest-auth==7.0.1
djangorestframework-simplejwt==5.3.1